typescriptifies LogMap
